# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Book Catalogue API",
    "description": "Book catalogue api consisting of authors, genre, publishers and books"
  },
  "host": "localhost:4242",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/genres/": {
      "get": {
        "tags": [
          "Genres"
        ],
        "description": "Get all genres",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "number",
            "required": false,
            "description": "page number"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "number",
            "required": false,
            "description": "number of genres returned"
          }
        ],
        "responses": {
          "200": {
            "description": "Genres fetched successfully",
            "schema": {
              "$ref": "#/definitions/genresResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Genres"
        ],
        "description": "Add a new genre",
        "parameters": [
          {
            "name": "genre",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AddGenre"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Genre added successfully",
            "schema": {
              "$ref": "#/definitions/addNewGenreResponse"
            }
          }
        }
      }
    },
    "/api/v1/genres/{genreId}": {
      "get": {
        "tags": [
          "Genres"
        ],
        "description": "Get a genre by id",
        "parameters": [
          {
            "name": "genreId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Genre fetched successfully",
            "schema": {
              "$ref": "#/definitions/genreResponse"
            }
          }
        }
      }
    },
    "/api/v1/genres/{genreId}/books": {
      "get": {
        "tags": [
          "Genres"
        ],
        "description": "Get books for a genre",
        "parameters": [
          {
            "name": "genreId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Books fetched successfully",
            "schema": {
              "$ref": "#/definitions/booksResponse"
            }
          }
        }
      }
    },
    "/api/v1/publishers/": {
      "get": {
        "tags": [
          "Publishers"
        ],
        "description": "Get all publishers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "number",
            "required": false,
            "description": "page number"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "number",
            "required": false,
            "description": "number of publishers returned"
          }
        ],
        "responses": {
          "200": {
            "description": "Books fetched successfully",
            "schema": {
              "$ref": "#/definitions/publishersResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Publishers"
        ],
        "description": "Add a new publisher",
        "parameters": [
          {
            "name": "publisher",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AddPublisher"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Publisher added successfully",
            "schema": {
              "$ref": "#/definitions/addNewPublisherResponse"
            }
          }
        }
      }
    },
    "/api/v1/publishers/{publisherId}": {
      "get": {
        "tags": [
          "Publishers"
        ],
        "description": "Get a publisher by id",
        "parameters": [
          {
            "name": "publisherId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Publisher fetched successfully",
            "schema": {
              "$ref": "#/definitions/publisherResponse"
            }
          }
        }
      }
    },
    "/api/v1/publishers/{publisherId}/books": {
      "get": {
        "tags": [
          "Publishers"
        ],
        "description": "Get books for a publisher",
        "parameters": [
          {
            "name": "publisherId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Books fetched successfully",
            "schema": {
              "$ref": "#/definitions/booksResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Genre": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Technology"
        }
      }
    },
...
    "Publisher": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Your avg publishers"
        }
      }
    },
...
    "genresResponse": {
      "type": "object",
      "properties": {
        "results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Genre"
          }
        },
        "total": {
          "type": "number",
          "example": 1343
        }
      }
    },
...
    "publishersResponse": {
      "type": "object",
      "properties": {
        "results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Publisher"
          }
        },
        "total": {
          "type": "number",
          "example": 33
        }
      }
    },
    "booksResponse": {
      "type": "object",
      "properties": {
        "results": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 1
              },
              "title": {
                "type": "string",
                "example": "Introduction to programming"
              },
              "description": {
                "type": "string",
                "example": ""
              },
              "price": {
                "type": "number",
                "example": 324
              },
              "pageCount": {
                "type": "number",
                "example": 522
              },
              "publishedDate": {
                "type": "string",
                "example": "23/03/2021"
              },
              "createdAt": {
                "type": "string",
                "example": "2021-10-11T16:29:19.321Z"
              },
              "updatedAt": {
                "type": "string",
                "example": "2021-10-11T16:29:19.321Z"
              },
              "genreId": {
                "type": "number",
                "example": 2
              }
            }
          }
        },
        "total": {
          "type": "number",
          "example": 324
        }
      }
    },
...
    "genreResponse": {
      "$ref": "#/definitions/Genre"
    },
...
    "publisherResponse": {
      "$ref": "#/definitions/Publisher"
    },
...
    "addNewPublisherResponse": {
      "$ref": "#/definitions/Publisher"
    },
    "addNewGenreResponse": {
      "$ref": "#/definitions/Genre"
    },
...
    "AddGenre": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Literature"
        }
      },
      "required": [
        "name"
      ]
    },
    "AddPublisher": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Your avg publishers"
        }
      },
      "required": [
        "name"
      ]
    }
  }
}